feat: agregar validaciones globales y constraints unique

+++ src/main/java/backend/Aplicacion/dto/puntaje/ObtenerPuntajesDTOResponse.java
@@ -2,10 +2,7 @@ package backend.Aplicacion.dto.puntaje;
 
 import java.util.List;
 
-
 public record ObtenerPuntajesDTOResponse(
         List<String> labels,
-        List<Integer> data
-
-) {
-}
+        List<Double> data
+) { }
